Remove submodules
Mbed TLS framework submodule is only needed for Mbed TLS development, it
is not required for vendoring. Since the source files of Mbed TLS are
added under src/ as a git subtree, it confuses the outer git repo and
fails the submodule update command that is executed by CI.

Remove the submodules to allow CI to pass.